intel: base_fw: uaol: move UAOL code out of base_fw
This moves UAOL-related code to a dedicated uaol.c file, which will be
extended with additional UAOL functionality in the future.
