+++ src/audio/CMakeLists.txt
@@ -126,6 +126,9 @@ if(NOT CONFIG_COMP_MODULE_SHARED_LIBRARY_BUILD)
 	if(CONFIG_COMP_TONE)
 		add_subdirectory(tone)
 	endif()
+	if(CONFIG_DAI_INTEL_UAOL)
+		add_local_sources(sof uaol.c)
+	endif()
 	if(CONFIG_ZEPHYR_NATIVE_DRIVERS)
 		list(APPEND base_files host-zephyr.c)
 		sof_list_append_ifdef(CONFIG_COMP_DAI base_files dai-zephyr.c)
