+++ src/audio/base_fw_intel.c
@@ -24,10 +24,6 @@
 #include <sof/lib/memory.h>
 #include <sof/lib_manager.h>
 
-#if CONFIG_UAOL_INTEL_ADSP
-#include <zephyr/drivers/uaol.h>
-#endif
-
 #include <ipc4/base_fw.h>
 #include <ipc4/alh.h>
 #include <rimage/sof/user/manifest.h>
@@ -35,6 +31,10 @@
 
 #if CONFIG_INTEL_ADSP_MIC_PRIVACY
 #include <sof/audio/mic_privacy_manager.h>
+#endif
+
+#if CONFIG_UAOL_INTEL_ADSP
+#include <sof/audio/uaol.h>
 #endif
 
 struct ipc4_modules_info {
@@ -45,22 +45,6 @@ struct ipc4_modules_info {
 /* Sanity check because a subtraction of those sizes is performed later on */
 STATIC_ASSERT(sizeof(struct ipc4_modules_info) < SOF_IPC_MSG_MAX_SIZE,
 	      invalid_modules_info_struct_size);
-
-#if CONFIG_UAOL_INTEL_ADSP
-struct ipc4_uaol_link_capabilities {
-	uint32_t input_streams_supported          : 4;
-	uint32_t output_streams_supported         : 4;
-	uint32_t bidirectional_streams_supported  : 5;
-	uint32_t rsvd                             : 19;
-	uint32_t max_tx_fifo_size;
-	uint32_t max_rx_fifo_size;
-} __packed __aligned(4);
-
-struct ipc4_uaol_capabilities {
-	uint32_t link_count;
-	struct ipc4_uaol_link_capabilities link_caps[];
-} __packed __aligned(4);
-#endif	/* CONFIG_UAOL_INTEL_ADSP */
 
 /*
  * TODO: default to value of ACE1.x platforms. This is defined
@@ -102,61 +86,6 @@ __cold int basefw_vendor_fw_config(uint32_t *data_offset, char *data)
 
 	return 0;
 }
-
-#if CONFIG_UAOL_INTEL_ADSP
-#define DEV_AND_COMMA(node) DEVICE_DT_GET(node),
-static const struct device *uaol_devs[] = {
-	DT_FOREACH_STATUS_OKAY(intel_adsp_uaol, DEV_AND_COMMA)
-};
-
-#if !CONFIG_SOF_OS_LINUX_COMPAT_PRIORITY
-__cold static void tlv_value_set_uaol_caps(struct sof_tlv *tuple, uint32_t type)
-{
-	const size_t dev_count = ARRAY_SIZE(uaol_devs);
-	struct uaol_capabilities dev_cap;
-	struct ipc4_uaol_capabilities *caps = (struct ipc4_uaol_capabilities *)tuple->value;
-	size_t caps_size = offsetof(struct ipc4_uaol_capabilities, link_caps[dev_count]);
-	size_t i;
-	int ret;
-
-	assert_can_be_cold();
-
-	memset(caps, 0, caps_size);
-
-	caps->link_count = dev_count;
-	for (i = 0; i < dev_count; i++) {
-		ret = uaol_get_capabilities(uaol_devs[i], &dev_cap);
-		if (ret)
-			continue;
-
-		caps->link_caps[i].input_streams_supported = dev_cap.input_streams;
-		caps->link_caps[i].output_streams_supported = dev_cap.output_streams;
-		caps->link_caps[i].bidirectional_streams_supported = dev_cap.bidirectional_streams;
-		caps->link_caps[i].max_tx_fifo_size = dev_cap.max_tx_fifo_size;
-		caps->link_caps[i].max_rx_fifo_size = dev_cap.max_rx_fifo_size;
-	}
-
-	tlv_value_set(tuple, type, caps_size, caps);
-}
-#endif /* CONFIG_SOF_OS_LINUX_COMPAT_PRIORITY */
-
-__cold static int uaol_stream_id_to_hda_link_stream_id(int uaol_stream_id)
-{
-	size_t dev_count = ARRAY_SIZE(uaol_devs);
-	size_t i;
-
-	assert_can_be_cold();
-
-	for (i = 0; i < dev_count; i++) {
-		int hda_link_stream_id = uaol_get_mapped_hda_link_stream_id(uaol_devs[i],
-									    uaol_stream_id);
-		if (hda_link_stream_id >= 0)
-			return hda_link_stream_id;
-	}
-
-	return -1;
-}
-#endif	/* CONFIG_UAOL_INTEL_ADSP */
 
 __cold int basefw_vendor_hw_config(uint32_t *data_offset, char *data)
 {
